Move tests for error cases
